+++ src/main/java/com/matchmetrics/MatchMetricsApplication.java
@@ -1,13 +1,13 @@
-package com.matchmomentummetrics;
+package com.matchmetrics;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class MatchMomentumMetricsApplication {
+public class MatchMetricsApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(MatchMomentumMetricsApplication.class, args);
+        SpringApplication.run(MatchMetricsApplication.class, args);
     }
 
 }
